move everyting from current issue to issues/issue-1
