+++ src/issues/issue-1/apartment-drawing.js
@@ -1,0 +1,12 @@
+import React from "react"
+import { PaddedWidthContainer } from "../../styles/layout"
+
+const ApartmentDrawing = () => {
+  return (
+    <PaddedWidthContainer>
+      <div>Aarprmtne Drawing</div>
+    </PaddedWidthContainer>
+  )
+}
+
+export default ApartmentDrawing
