+++ src/issues/issue-1/index.js
@@ -1,0 +1,15 @@
+import React from "react"
+
+import Cover from "./cover"
+import ApartmentDrawing from "./apartment-drawing"
+
+const Issue1 = () => {
+  return (
+    <>
+      <Cover />
+      <ApartmentDrawing />
+    </>
+  )
+}
+
+export default Issue1
