+++ src/styles/global.css
@@ -16,21 +16,21 @@ body {
 
 h1 {
   font-family: "HelveticaNeue-Bold";
-  font-size: 73px;
-  color: #ff07a9;
-  margin: 0;
-  line-height: 70px;
+  font-size: 55px;
+  line-height: 65px;
+  color: var(--pink);
 }
 
 h2 {
   font-family: "HelveticaNeue-Medium";
-  font-size: 67px;
-  color: #ff07a9;
+  font-size: 50px;
+  line-height: 50px;
+  color: var(--pink);
 }
 
 p,
 a {
   font-family: "HelveticaNeue-Medium";
-  font-size: 16px;
+  font-size: 30px;
   color: var(--black);
 }
