+++ src/styles/layout.js
@@ -1,5 +1,8 @@
 import styled from "styled-components"
 
 export const PaddedWidthContainer = styled.section`
-  padding: 0 145px;
+  padding: 145px;
+`
+export const FullWidthContainer = styled.section`
+  padding: 145px 0;
 `
